docs/requirements.txt
furo
sphinx-design
setuptools-rust
Pygments